Upgrade parse_freq(), parse_time() to return double
Hope that doesn't break anything

+++ include/units.h
@@ -1,6 +1,6 @@
 
-uint32_t parse_freq(double freq, const char* unit);
-uint32_t parse_time(double time, const char* unit);
+double parse_freq(double freq, const char* unit);
+double parse_time(double time, const char* unit);
 double parse_volts(double voltage, const char* unit);
 char* time_unit(double time);
 char* freq_unit(double freq);
